Pre-load Mart
The marts for human / Mouse are saved, and are loaded at the same time as the data to cut loading time & internet need. Will do in the future a button to update the mart for selected species.
For now it checks if the mart exist, and if not, it queries and saves it.